BP5: 32<->64bit round-trip
Explicitly ok to truncate data that is >4GB on 32bit systems.

+++ source/adios2/toolkit/format/bp5/BP5Deserializer.h
@@ -163,8 +163,8 @@ private:
         char *VarName = NULL;
         size_t DimCount = 0;
         size_t JoinedDimen = SIZE_MAX;
-        size_t *LastJoinedOffset = NULL;
-        size_t *LastJoinedShape = NULL;
+        uint64_t *LastJoinedOffset = NULL;
+        uint64_t *LastJoinedShape = NULL;
         bool Derived = false;
         char *ExprStr = NULL;
         ShapeID OrigShapeID;
@@ -174,7 +174,7 @@ private:
         DataType Type;
         int ElementSize = 0;
         size_t MinMaxOffset = SIZE_MAX;
-        size_t *GlobalDims = NULL;
+        uint64_t *GlobalDims = NULL;
         size_t LastTSAdded = SIZE_MAX;
         size_t FirstTSSeen = SIZE_MAX;
         size_t LastStepAdded = SIZE_MAX;
@@ -245,7 +245,7 @@ private:
     std::vector<std::vector<void *> *> MetadataBaseArray;
     // for random access mode, for each timestep, for each writerrank, base
     // address of the joined dim arrays, for streaming use 0 index
-    std::vector<std::vector<size_t *>> JoinedDimArray;
+    std::vector<std::vector<uint64_t *>> JoinedDimArray;
     size_t JDAIdx = 0;
 
     ControlInfo *ControlBlocks = nullptr;
@@ -257,7 +257,7 @@ private:
     BP5VarRec *LookupVarByKey(void *Key) const;
     BP5VarRec *LookupVarByName(const char *Name);
     BP5VarRec *CreateVarRec(const char *ArrayName);
-    void ReverseDimensions(size_t *Dimensions, size_t count, size_t times);
+    void ReverseDimensions(uint64_t *Dimensions, size_t count, size_t times);
     const char *BreakdownVarName(const char *Name, DataType *type_p, int *element_size_p);
     void BreakdownFieldType(const char *FieldType, bool &Operator, bool &MinMax);
     void BreakdownArrayName(const char *Name, char **base_name_p, DataType *type_p,
@@ -266,7 +266,7 @@ private:
                               int *element_size_p, bool &Operator, bool &MinMax);
     void *VarSetup(core::Engine *engine, const char *variableName, const DataType type, void *data);
     void *ArrayVarSetup(core::Engine *engine, const char *variableName, const DataType type,
-                        int DimCount, size_t *Shape, size_t *Start, size_t *Count,
+                        int DimCount, uint64_t *Shape, uint64_t *Start, uint64_t *Count,
                         core::StructDefinition *Def, core::StructDefinition *ReaderDef);
     void MapGlobalToLocalIndex(size_t Dims, const size_t *GlobalIndex, const size_t *LocalOffsets,
                                size_t *LocalIndex);
@@ -281,7 +281,7 @@ private:
     void StructQueueReadChecks(core::VariableStruct *variable, BP5VarRec *VarRec);
 
     void *GetMetadataBase(BP5VarRec *VarRec, size_t Step, size_t WriterRank) const;
-    bool IsContiguousTransfer(BP5ArrayRequest *Req, size_t *offsets, size_t *count);
+    bool IsContiguousTransfer(BP5ArrayRequest *Req, uint64_t *offsets, uint64_t *count);
     char *FillBlock(std::map<BP5VarRec *, MinVarInfo *> &map);
 
     size_t CurTimestep = 0;
